Added Travis integration
This is how I originally set up Travis integration for our GitHub repos. It's based off of [R Travis](https://github.com/craigcitro/r-travis/). So the step involved are as such:

Add a `.travis.yml` file to the repo. Here we establish what needs to be installed. It basically installs all dependencies assuming they are available on CRAN. You can also install packages that exist only on GitHub like so:

```
- ./travis-tool.sh github_package assertthat
```

Note: We no longer have to do this for `assertthat`.

I also created an encrypted API key to add to the yml file so we have notifications in the Package Development room of our HipChat. Usually build pass/fail notifications go there. Especially useful if you want to turn off extraneous email notifications.

Then we have to turn on Travis integration on [Travis CI](https://travis-ci.org/). I already did that for rEML now but you can do it too. I think if you just sign on with your GitHub account, it shows all repos and organizations you have access to.

Finally, from the Travis page for a repo, you can grab the build shield from the top right of the screen. I've added it to the README.