fixed some citation definitions and tests Coercion to citation should be of type citation, not book, editedBook etc Also re-defined the EML citation class as Citation (capital C), to avoid conflict with the base R 'citation' class.
Since this change means the class name does not match the XML element name,
I added a mechanism to S4Toeml using a custom slot, xmlNodeName.  For emlToS4
this is already handled by specifying the class name explicitly in the function
call.  This may be useful for other classes as well