hack element ordering Introduces an imprecise hack for element ordering.  Previously, xs:goup elements like 'ResourceGroup' were inherited via 'contains'.  However, serializing these elements follows the slot order, and slots inherited via contains are always listed after slots assigned with 'slots' directly.
To address this, we now parse xs:group into a slot with class xs:group[@ref]. The
previous R code serialized this as an explicit element, e.g. <ReferenceGroup>, and
also did not know how to parse this, since instead of the list of inherited slots,
we now had a single slot the inherited class.  Thus, emlToS4 and S4Toeml have now
been modified to handle this special case, by treating any slot of with class that
begins by being capitalized (a convention of the EML Schema, but a bit risky) as
a 'metaslot', e.g. treating it's children as if they were slots of the parent class.