Refactored to allow validation against multiple EML versions.
The prior release assumed that all documents should be validated against the EML 2.1.1 schema.  With this change, the XML namespace for the root element is used to determine the schema to be used for validation, which prevents the schema not found error described in issue #185.  To accomodate this, the current EML schema files were moved into the `inst/xsd/eml-2.1.1` directory, and a new directory was created for schema files from `eml-2.1.0`.  Still need to add files and tests for prior versions of EML.

I added a dependency on the `stringr` package -- let me know @cboettig if that is ok.