Update package to use EML 2.2.0 constructors
Addresses #287. Note that dynamically switching the constructors based on version isn't supported here: We just have constructors for a single version (2.2.0) at a time.

Note this depends on https://github.com/ropensci/emld/pull/66 landing in master to work right because we build the constructors from the copy of the JSON file in master.