Fix #311
The standardUnits object created in standardUnits.R and saved to sysdata.rda appears to be inaccurate. This fix refreshes the standardUnits object and sysdata.rda. Additionally, the udunits_units object was not formally saved to sysdata.rda by package code, but now is. Furthermore, the deprecated devtools::use_data() has been replaced by usethis::use_data() and the argument overwrite set to TRUE.