Fix three URLs to satisfy CRAN checks
Note that this breaks the DOI in the creating-EML vignette because the PNAS DOI 503s without either JS or cookies enabled (not sure which) due to that URL being fronted with Cloudflare and probably some sort of anti-abuse config PNAS has with them. This is regrettable but being on CRAN is better than making DOIs URLs.